EIP1271VerifierTests unit tests
