+++ Example/IntegrationTests/Chat/RegistryTests.swift
@@ -7,7 +7,7 @@ import WalletConnectUtils
 final class RegistryTests: XCTestCase {
 
     func testRegistry() async throws {
-        let client = HTTPClient(host: "keys.walletconnect.com")
+        let client = HTTPNetworkClient(host: "keys.walletconnect.com")
         let registry = KeyserverRegistryProvider(client: client)
         let account = Account("eip155:1:" + Data.randomBytes(count: 16).toHexString())!
         let pubKey = SigningPrivateKey().publicKey.hexRepresentation
