+++ Sources/Auth/Services/Signer/MessageSignerFactory.swift
@@ -13,7 +13,7 @@ public struct MessageSignerFactory {
             eip191Verifier: EIP191Verifier(),
             eip1271Verifier: EIP1271Verifier(
                 projectId: projectId,
-                httpClient: HTTPClient(host: "rpc.walletconnect.com")
+                httpClient: HTTPNetworkClient(host: "rpc.walletconnect.com")
             )
         )
     }
